requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
black
flake8
isort
autopep8
google-generativeai
python-dotenv
requests
